+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## v2.0.11 (1.21)
 - improved enchant compat on neoforge
+- re-added armor-stand to workbench UI
 - fixed breeze rod giving immolate
 - fixed on-hit-potions for modular arrows
 - fixed language related crash
 - fixed enchantments not applying on area harvest
 - fixed ht treechop issues on neoforge
 - fixed issues with Refined Storage
+- fixed mining level behaviour